+++ dms-enterprise-20181101/ChangeLog.md
@@ -1,3 +1,8 @@
+2025-02-13 Version: 1.66.1
+- Update API UpdateDataLakeTable: add param TableName.
+- Update API UpdateUser: add param UidString.
+
+
 2025-02-11 Version: 1.66.0
 - Support API CreateDifyInstance.
 - Support API DescribeDifyDefaultVpc.